No reason to mess with the Kafka setup when testing initial timeout
Note also that due to topic naming mistakes the partitioning from
the timeout test affected the run test
